+++ src/utils.rs
@@ -433,7 +433,8 @@ pub mod test_utils {
     pub(crate) use check_scope;
 
     macro_rules! scope {
-        (  $( ($name: expr, $val: expr)),*  ) => {
+        () => { ExecutionScopes::new() };
+        (  $( ($name: expr, $val: expr)),* $(,)?  ) => {
             {
                 let mut exec_scopes = ExecutionScopes::new();
                 $(