+++ CHANGELOG.md
@@ -34,6 +34,21 @@
         %}
     ```
 
+* Add missing hint on cairo_secp lib [#1089](https://github.com/lambdaclass/cairo-rs/pull/1089):
+    `BuiltinHintProcessor` now supports the following hint:
+
+    ```python
+
+    from starkware.cairo.common.cairo_secp.secp_utils import pack
+
+    slope = pack(ids.slope, PRIME)
+    x0 = pack(ids.point0.x, PRIME)
+    x1 = pack(ids.point1.x, PRIME)
+    y0 = pack(ids.point0.y, PRIME)
+
+    value = new_x = (pow(slope, 2, SECP_P) - x0 - x1) % SECP_P
+    ```
+
 * Add missing hint on vrf.json whitelist [#1055](https://github.com/lambdaclass/cairo-rs/pull/1055):
 
      `BuiltinHintProcessor` now supports the following hint:
@@ -137,6 +152,31 @@
         x = pack(ids.x, PRIME) % SECP_P
     %}
 
+* Add missing hints `NewHint#55`, `NewHint#56`, and `NewHint#57` [#1077](https://github.com/lambdaclass/cairo-rs/issues/1077)
+
+    `BuiltinHintProcessor` now supports the following hints:
+
+    ```python
+    from starkware.cairo.common.cairo_secp.secp_utils import pack
+    SECP_P=2**255-19
+
+    x = pack(ids.x, PRIME) % SECP_P
+    ```
+
+    ```python
+    from starkware.cairo.common.cairo_secp.secp_utils import pack
+    SECP_P=2**255-19
+
+    value = pack(ids.x, PRIME) % SECP_P
+    ```
+
+    ```python
+    SECP_P=2**255-19
+    from starkware.python.math_utils import div_mod
+
+    value = x_inv = div_mod(1, x, SECP_P)
+    ```
+    
 * Implement hint for `starkware.cairo.common.cairo_keccak.keccak._copy_inputs` as described by whitelist `starknet/security/whitelists/cairo_keccak.json` [#1058](https://github.com/lambdaclass/cairo-rs/pull/1058)
 
     `BuiltinHintProcessor` now supports the following hint:
@@ -734,6 +774,7 @@
             if root_gx == None:
                 root_gx = 0
             ids.success_x = int(success_x)
+            ids.success_gx = int(success_gx)
             split_root_x = split(root_x)
             split_root_gx = split(root_gx)
             ids.sqrt_x.d0 = split_root_x[0]